+++ packages/sdk/src/sync/subgraph/fetch.ts
@@ -6,6 +6,10 @@ import {
   IncludedNote,
 } from "../../primitives";
 import { maxArray } from "../../utils";
+import {
+  entityIdWithEntityIndexFromBlockNumber,
+  makeSubgraphQuery,
+} from "./utils";
 
 interface NoteResponse {
   ownerH1X: string;
@@ -45,13 +49,13 @@ interface FetchNotesResponse {
 }
 
 interface FetchNotesVars {
-  fromBlock: number;
-  toBlock: number;
+  fromID: string;
+  toID: string;
 }
 
 const notesQuery = `\
-query fetchNotes($fromBlock: Int!, $toBlock: Int!) {
-  encodedOrEncryptedNotes(block: { number: $toBlock, number_gte: $fromBlock }) {
+query fetchNotes($fromID: Bytes!, $toID: Bytes!) {
+  encodedOrEncryptedNotes(where: { id_gte: $fromID, id_lt: $toID}) {
     merkleIndex
     note {
       ownerH1X
@@ -90,7 +94,11 @@ export async function fetchNotes(
     notesQuery,
     "notes"
   );
-  const res = await query({ fromBlock, toBlock });
+
+  const fromID = entityIdWithEntityIndexFromBlockNumber(BigInt(fromBlock));
+  const toID = entityIdWithEntityIndexFromBlockNumber(BigInt(toBlock + 1));
+
+  const res = await query({ fromID, toID });
   return res.data.encodedOrEncryptedNotes.map(
     ({ merkleIndex, note, encryptedNote }) => {
       if (note) {
@@ -228,8 +236,8 @@ interface FetchNullifiersResponse {
 }
 
 interface FetchNullifiersVars {
-  fromBlock: number;
-  toBlock: number;
+  fromID: string;
+  toID: string;
 }
 
 interface NullifierResponse {
@@ -237,8 +245,8 @@ interface NullifierResponse {
 }
 
 const nullifiersQuery = `
-  query fetchNullifiers($fromBlock: Int!, $toBlock: Int!) {
-    nullifiers(block: { number: $toBlock, number_gte: $fromBlock}) {
+  query fetchNullifiers($fromID: Bytes!, $toID: Bytes!) {
+    nullifiers(where: { id_gte: $fromID, id_lt: $toID}) {
       nullifier
     }
   }
@@ -256,69 +264,10 @@ export async function fetchNullifiers(
     nullifiersQuery,
     "nullifiers"
   );
-  const res = await query({ fromBlock, toBlock });
+
+  const fromID = entityIdWithEntityIndexFromBlockNumber(BigInt(fromBlock));
+  const toID = entityIdWithEntityIndexFromBlockNumber(BigInt(toBlock + 1));
+
+  const res = await query({ fromID, toID });
   return res.data.nullifiers.map(({ nullifier }) => BigInt(nullifier));
 }
-
-// see https://thegraph.com/docs/en/querying/graphql-api/#subgraph-metadata
-const latestIndexedBlockQuery = `
-{
-  _meta {
-    block {
-      number
-    }
-  }
-}
-`;
-
-interface FetchLatestIndexedBlockResponse {
-  data: {
-    _meta: {
-      block: {
-        number: number;
-      };
-    };
-  };
-}
-
-// gets the latest indexed block from the subgraph
-export async function fetchLatestIndexedBlock(
-  endpoint: string
-): Promise<number> {
-  const query = makeSubgraphQuery<undefined, FetchLatestIndexedBlockResponse>(
-    endpoint,
-    latestIndexedBlockQuery,
-    "latest indexed block"
-  );
-  const res = await query(undefined);
-  return res.data._meta.block.number;
-}
-
-export const makeSubgraphQuery =
-  <T, U>(endpoint: string, query: string, dataLabel: string) =>
-  async (variables: T): Promise<U> => {
-    try {
-      const response = await fetch(endpoint, {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify({
-          query,
-          variables,
-        }),
-      });
-
-      if (!response.ok) {
-        const text = await response.text();
-        console.error(`Failed to query ${dataLabel} from subgraph: ${text}`);
-
-        throw new Error(`Failed to query ${dataLabel} from subgraph: ${text}`);
-      }
-
-      return (await response.json()) as U;
-    } catch (err) {
-      console.error(`Error when querying ${dataLabel} from subgraph`);
-      throw err;
-    }
-  };