+++ packages/sdk/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+- move shared subgraph utils to their own module
+- fix `SubgraphSDKSyncAdapter` querying entire history instead of only specified range
 - Remove `chainId` from `DepositRequest` (already included in eip712 domain)
 - Make syncing rely on `handler` contract after contract separation work
 - Export `IterSync` opts for `deposit-screener` package to use as well